pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dlive-midi-bridge"
version = "0.1.0"
description = "RTP-MIDI to Allen & Heath dLive TCP bridge — eliminate the Mac Mini"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Backline Logic", email = "michael@backlinelogic.com"},
]
keywords = ["midi", "dlive", "allen-heath", "rtp-midi", "live-sound"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Sound/Audio",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]

dependencies = [
    "zeroconf>=0.131.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest", "black", "ruff"]

[project.scripts]
dlive-midi-bridge = "dlive_midi_bridge.cli:main"
dlive-test-send = "dlive_midi_bridge.test_send:main"

[tool.setuptools.packages.find]
where = ["src"]
